# microservice-course/src/main/java/com/example/course/entity/Section.java
package com.example.course.entity;

import com.baomidou.mybatisplus.annotation.IdType;
import com.baomidou.mybatisplus.annotation.TableId;
import com.baomidou.mybatisplus.annotation.TableName;
import lombok.Data;

@Data
@TableName("section")
public class Section {
    @TableId(type = IdType.AUTO)
    private Long id;
    private Long chapterId;
    private String title;
    private String content;
    private String videoUrl;
    private Integer duration;
    private Integer sort;
    private Boolean isFree;
}